requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.5
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.1
rich>=13.7.1
typer>=0.12.3
chromadb>=0.5.5
sentence-transformers>=3.0.1
numpy>=1.26.4
requests>=2.32.3
reportlab>=4.2.0
